Relax httpoison dependency
Currently we depend on httpoison ~> 0.7, which requires httpoison < 1.0.0.
This can prevent using the SDK in a modern application, since httpoison 1.0.0
was released in 2018, and httpoison 2.0.0 was released in 2023.

Specifying "~> 0.7 or ~> 1.0 or ~> 2.0" allows an app to use whatever version of
httpoison might be required by other dependencies, and still be compatible with
this SDK.